+++ packages/appcd-util/CHANGELOG.md
@@ -1,6 +1,7 @@
 # v3.1.0
 
  * feat: Added `redact()` function to scrub sensitive information from a value.
+ * feat: Added re-export for Lodash's `set()` function.
 
 # v3.0.0 (Jun 12, 2020)